requirements.txt
requests==2.32.3
loadotenv==1.0.1
SQLAlchemy==2.0.38
pytest==8.3.4
fastapi==0.115.10
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
python-dateutil
psycopg2-binary
alembic==1.14.1
pytest-asyncio==0.25.3
httpx